Rename exit survey file
